==> code/analysis_code/1_EDA_sex.Rmd <==
```{r message=FALSE}
library(here)
library(GGally)
library(skimr)
library(ggfortify)
library(MASS)
library(tidyverse)
library(readxl)
```
Loading data
```{r}
sex_df <- readRDS(here("data/processed_data/sex_df_cleaned.rds"))
```

Creating `Expected Death Rate` that will be used to calculate the standardized mortality ratio. Filter for only states in the contiguous United States.
```{r}
sex_df <- sex_df %>% 
 mutate(`Expected Death Rate` = (`Expected Death Cases`/Population)*1e5,
        `Standardized Mortality Ratio` = `Death Rate (estimated)`/(`Expected Death Rate`)) %>%
  filter(!grepl("Alaska", State),
         !grepl("Hawaii", State))
```
Carrying out some diagnostic plots on key variables. 
```{r}
autoplot(lm(sex_df$`Death Rate (AIDSVu)` ~ 1))
autoplot(lm(sex_df$`Death Rate (estimated)` ~ 1))
autoplot(lm(sex_df$`Ratio Observed/Expected` ~ 1))
autoplot(lm(sex_df$`Standardized Mortality Ratio` ~ 1))
```

```{r}
summary(sex_df$`Death Rate (AIDSVu)`)
summary(sex_df$`Death Rate (estimated)`)
summary(sex_df$`Ratio Observed/Expected`)
summary(sex_df$`Standardized Mortality Ratio`)
```
It appears that our data does not follow a normal distribution. Let's start by obtaining a histogram.
```{r}
sex_df %>% 
  ggplot() +
  geom_histogram(aes(x = `Death Rate (estimated)`, fill = Sex), bins = 10)
shapiro.test(sex_df$`Death Rate (estimated)`)

sex_df %>% 
  ggplot() +
  geom_histogram(aes(x = `Ratio Observed/Expected`, fill = Sex), bins = 10)
shapiro.test(sex_df$`Ratio Observed/Expected`)

sex_df %>% 
  ggplot() +
  geom_histogram(aes(x = `Standardized Mortality Ratio`, fill = Sex), bins = 10)
shapiro.test(sex_df$`Standardized Mortality Ratio`)
```

Visual assessment and the sharpiro test are in agreement that the data is not normally distributed. Let's see if log (x +1) transformation can help resolve this issue.

```{r}
sex_df %>% 
  ggplot() +
  geom_histogram(aes(x = log(`Death Rate (estimated)` + 1), fill = Sex) , bins = 10)
shapiro.test(log(sex_df$`Death Rate (estimated)` + 1))

sex_df %>% 
  ggplot() +
  geom_histogram(aes(x = log(`Ratio Observed/Expected` + 1), fill = Sex) , bins = 10)
shapiro.test(log(sex_df$`Ratio Observed/Expected` + 1))

sex_df %>% 
  ggplot() +
  geom_histogram(aes(x = log(`Standardized Mortality Ratio` +1), fill = Sex) , bins = 10)
shapiro.test(log(sex_df$`Standardized Mortality Ratio` + 1))
```

While this is much better, we found sufficient evidence to reject that the distribution is normal with p values < 0.05. Let's see if another transformation method can achieve better results. Other options include power transformations, like Box Cox transformation or Yeo Johnson transformation, and mixture models, such as the zero-inflated model (Poisson or negative binomial).

I'll try the Box-Cox transformation first. A limitation of this is that Box Cox cannot values of zero (or negative numbers). We will have to subset observations greater than zero to continue with the transformation. We will use boxcox() to find lambda. 
`library(MASS)`
```{r}
sex_df_DR <- sex_df %>% 
  filter(`Death Rate (estimated)` > 0)
b_DR <-  boxcox(lm(sex_df_DR$`Death Rate (estimated)` ~ 1))
lambda_DR <- b_DR$x[which.max(b_DR$y)]

sex_df_ROE <- sex_df %>% 
  filter(`Ratio Observed/Expected` > 0)

b_ROE <-  boxcox(lm(sex_df_ROE$`Ratio Observed/Expected` ~ 1))
lambda_ROE <- b_ROE$x[which.max(b_ROE$y)]

sex_df_SMR <- sex_df %>% 
  filter(`Standardized Mortality Ratio`  > 0)
b_SMR <-  boxcox(lm(sex_df_SMR$`Standardized Mortality Ratio` ~ 1))
lambda_SMR <- b_SMR$x[which.max(b_SMR$y)]

sex_df_SMR <- sex_df %>% 
  filter(`Standardized Mortality Ratio`  > 0)

```

We'll create a function to carry out the box cox transformation [cite](https://stats.stackexchange.com/questions/1444/how-should-i-transform-non-negative-data-including-zeros)
```{r}
box_cox <- function(x, lambda) {

    eps <- 0.00001
    if (abs(lambda) < eps)
        log(x)
    else
        (x ^ lambda - 1) / lambda

}
```

Box Cox transformation of DR and results of shapiro test.
```{r}
sex_df_DR %>%
  mutate(BOX = box_cox(`Death Rate (estimated)`, lambda_DR)) %>%
  ggplot() + geom_histogram(aes(x= BOX, fill = Sex), bins = 10)

sex_df_DR <- sex_df_DR %>%
  mutate(BOX = box_cox(`Death Rate (estimated)`, lambda_DR))

sex_df_DR %>%
  pull(BOX) %>% 
  shapiro.test()

sex_df_DR %>%
  filter(Sex == "Female") %>% 
  pull(BOX) %>% 
  shapiro.test()

sex_df_DR %>%
  filter(Sex == "Male") %>% 
  pull(BOX) %>% 
  shapiro.test()

par(mfrow=c(1,3))
plot(lm(sex_df_DR$BOX ~ 1))
```
Unsuccessful at normalizing distribution of Death Rate, and overall worse results than log transformation, except for the males.

Box Cox transformation of ROE and results of shapiro test.
```{r}
sex_df_ROE %>%
  mutate(BOX = box_cox(`Ratio Observed/Expected`, lambda_ROE)) %>%
  ggplot() + geom_histogram(aes(x= BOX, fill = Sex), bins = 10)

sex_df_ROE <- sex_df_ROE %>%
  mutate(BOX = box_cox(`Ratio Observed/Expected`, lambda_ROE))

sex_df_ROE %>%
  pull(BOX) %>% 
  shapiro.test()

sex_df_ROE %>%
  filter(Sex == "Female") %>% 
  pull(BOX) %>% 
  shapiro.test()

sex_df_ROE %>%
  filter(Sex == "Male") %>% 
  pull(BOX) %>% 
  shapiro.test()

par(mfrow=c(1,3))
plot(lm(sex_df_ROE$BOX ~ 1))
```
The Box Cox transformation was able to normalize the `Ratio Observed/Expected`. We will use the Box Cox transformation method for analyses with `Ratio Observed/Expected`. 

Box Cox transformation of SMR and results of shapiro test.
```{r}
sex_df_SMR %>%
  mutate(BOX = box_cox(`Standardized Mortality Ratio`, lambda_SMR)) %>%
  ggplot() + geom_histogram(aes(x= BOX, fill = Sex), bins =10) 

sex_df_SMR <- sex_df_SMR %>%
  mutate(BOX = box_cox(`Standardized Mortality Ratio`, lambda_SMR))

sex_df_SMR %>%
  filter(Sex == "Female") %>% 
  pull(BOX) %>% 
  shapiro.test()

sex_df_SMR %>%
  filter(Sex == "Male") %>% 
  pull(BOX) %>% 
  shapiro.test()

par(mfrow=c(1,3))
plot(lm(sex_df_SMR$BOX ~ 1))
```
The Box Cox transformation was not technically able to normalize the `Standardized Mortality Ratio`. We will move on the Yeo Johnson transformation


We will perform Yeo Johnson transformation on `Death Rate (estimated)` and `Standardized Mortality Ratio`.We will use `yeojohnson` from `BestNormalize` package to extract lambda and `yeo_johnson` to compute the transformation for each observation. Code for `yeo_johnson` obtained from [GitHub](https://stats.stackexchange.com/questions/1444/how-should-i-transform-non-negative-data-including-zeros)
```{r}
library(bestNormalize)

yeo_johnson <- function(x, lambda) {

    eps <- .000001
    not_neg <- which(x >= 0)
    is_neg  <- which(x < 0)

    not_neg_trans <- function(x, lambda) {
        if (abs(lambda) < eps) log(x + 1)
        else ((x + 1) ^ lambda - 1) / lambda
    }

    neg_trans <- function(x, lambda) {
        if (abs(lambda - 2) < eps) - log(-x + 1)
        else - ((-x + 1) ^ (2 - lambda) - 1) / (2 - lambda)
    }

    x[not_neg] <- not_neg_trans(x[not_neg], lambda)

    x[is_neg] <- neg_trans(x[is_neg], lambda)

    return(x)

}

```

```{r}
yj_DR <- yeojohnson(sex_df_DR$`Death Rate (estimated)`)
yj_DR <- yj_DR$lambda

sex_df <- sex_df %>%
  mutate(YJ = yeo_johnson(`Death Rate (estimated)`, yj_DR))

sex_df %>%
  ggplot() + geom_histogram(aes(x= YJ, fill = Sex), bins =10) 

sex_df %>%
  pull(YJ) %>% 
  shapiro.test()

sex_df %>%
  filter(Sex == "Female") %>% 
  pull(YJ) %>% 
  shapiro.test()

sex_df %>%
  filter(Sex == "Male") %>% 
  pull(YJ) %>% 
  shapiro.test()

par(mfrow=c(1,3))
plot(lm(sex_df$YJ ~ 1))
```
Unable to normalized the distribution `Death Rate (estimated)` with the Yeo Johnson transformation, but improved for overall and males compared to log transformation (not for females).

```{r}
yj_SMR <- yeojohnson(sex_df_SMR$`Standardized Mortality Ratio`)
yj_SMR <- yj_SMR$lambda

sex_df_SMR <- sex_df_SMR%>%
  mutate(YJ = yeo_johnson(`Standardized Mortality Ratio`, yj_SMR))

sex_df_SMR %>%
  ggplot() + geom_histogram(aes(x= YJ, fill = Sex), bins =10) 

sex_df_SMR %>%
  pull(YJ) %>% 
  shapiro.test()

sex_df_SMR %>%
  filter(Sex == "Female") %>% 
  pull(YJ) %>% 
  shapiro.test()

sex_df_SMR %>%
  filter(Sex == "Male") %>% 
  pull(YJ) %>% 
  shapiro.test()

par(mfrow=c(1,3))
plot(lm(sex_df_SMR$YJ ~ 1))
```
The Yeo Johnson transformation was only able to normalize the SMR for male, but not females or for both sexes combined.

Plotting correlation matrix
```{r message=FALSE, warning=FALSE}
X <- sex_df[,c(9:10, 16:25)]
ggpairs(X)
```
Let's save the dataframe for ROE since this one was able to be normalized after Box Cox transformation.
```{r}
saveRDS(sex_df_ROE, here("data/processed_data/sex_new.rds"))
```
