==> code/analysis_code/2_pca_alt.Rmd <==
```{r message=FALSE, warning=FALSE}
library(ggcorrplot)
library(FactoMineR)
library(factoextra)
library(here)
library(knitr)
library(tidyverse)
library(broom)
library(pls)

race_sub <- readRDS(here("data/processed_data/race_alt.rds"))
race_sub <- race_sub %>% ungroup()
pos_predictors <- readRDS(here("data/processed_data/pos_predictors.rds"))
sub <-race_sub %>% 
  select(contains(c("State", "Race", "Ratio")), matches(pos_predictors))
```

PCA
```{r}
#Setting up correlation matrix
pca_1 <- race_sub %>% 
  ungroup() %>% 
  select(all_of(pos_predictors)) %>% 
  select(-contains("Late")) %>%  ##poorly correlated
  cor() %>% 
  ggcorrplot()

# selects variables of interest and scale
pca <- race_sub %>% 
  select(matches(pos_predictors)) %>% 
  select(-contains("Late")) %>% 
  princomp(cor = TRUE)

```

```{r}
print(summary(pca))

#Scree plot
pca_2 <-print(fviz_eig(pca, addlabels = TRUE))

#pca loadings of principal components 1 and 2
print(pca$loadings[, 1:2])

#Storing PCA scores to use for fitting model
prin_comp1 <- pca$scores[,1]

## save
saveRDS(prin_comp1, here("data/processed_data/prin_comp1.rds"))
```

Since we originally had 6 variables, we should only look at components that (individually) explain more than 16.67% of variation. This is includes principal component 1 and 2. All other principal components will be excluded. 

```{r}
#Contribution of each variable 
pca_3 <-print(fviz_cos2(pca, choice = "var", axes = 1:2))

# Biplot of the attributes
pca_4 <- print(fviz_pca_var(pca, col.var = "cos2",
                   gradient.cols = c("darkblue", "orange", "green"),
                   repel = TRUE, title = "PCA"
                   ))
# Variables in green are a good representation of the variable on that component.
```

```{r}
library(gridExtra)
grid_pca <- grid.arrange(pca_1, pca_2, pca_3, pca_4, ncol = 2, nrow = 2)

ggsave(here("results/grid_pca.png"), grid_pca, units="in", width=10, height=10, dpi=300)
```
